backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==3.2.18
django-admin-honeypot==1.1.0
django-allauth==0.54.0
django-ipware==5.0.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
zipp==3.15.0
